refactor(admin-service): moves permission code from handler to permission middleware file
