+++ services/admin/backend/api/handler/BUILD.bazel
@@ -9,6 +9,7 @@ go_library(
     importpath = "github.com/dasch-swiss/dasch-service-platform/services/admin/backend/api/handler",
     visibility = ["//visibility:public"],
     deps = [
+        "//services/admin/backend/api/middleware",
         "//services/admin/backend/api/presenter",
         "//services/admin/backend/entity",
         "//services/admin/backend/entity/organization",
@@ -18,6 +19,5 @@ go_library(
         "//shared/go/pkg/valueobject",
         "@com_github_gorilla_mux//:go_default_library",
         "@com_github_urfave_negroni//:go_default_library",
-        "@com_github_golang_jwt_jwt//:go_default_library",
     ],
 )
